Fix building against older SDKs.
`NSConditionalExpressionType` and related functionality is new in OS X
10.11 and iOS 9. When building against an older SDK, provide our own
declarations for the constant and related interfaces on
`NSExpression`. This allows Realm to handle the newer expression type
which could in theory happen when used in an app targeting a newer OS
version when the Realm framework was built against an older SDK version.